service/mvd/requirements.txt
# conda install swig=3.0
seaborn
pytesseract
streamlit
textract
flask
jamspell
opencv-python
python-docx